refactor #1: Store를 생성하는 createStore 함수 생성
- initialState와 state를 변경하는 reducer 함수를 인자로 받는다.
- 현재 state를 반환하는 getState, action을 전달하는 dispatch, state의 변화에 따라 재실행되는(state를 구독하는) 함수를 등록하는 register 함수를 반환한다.
